% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{acceleration}
\alias{acceleration}
\title{A function to calculate the acceleration using velocities obtained based on interval of 10 seconds.}
\usage{
acceleration(trip)
}
\arguments{
\item{trip}{The trip of the driver}
}
\value{
It will return the acceleration of the trip in km/h based on 10 second intervals.
}
\description{
Based on a certain trip specified, this function will be able to calculate the acceleration using
the velocities of the trip that is based on interval of 10 seconds.
}
\examples{
# Acceleration of the trip (based on velocity10)
 acceleration(trip=sampleTrip)
}

