% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Driver.R
\docType{class}
\name{Driver}
\alias{Driver}
\title{R6 class for Driver}
\format{A \code{\link{R6Class}} generator object}
\usage{
Driver
}
\description{
This is a R6 class for Driver which contains methods to obtain various plots and information
about the driver.
}
\section{Methods}{

\describe{
 \item{\code{ShowTrip(trip)}}{This method uses \code{trip} as an argument to return a plot of 4 features (X-Y coordinates, velocity, acceleration, breaks).}
 \item{\code{ShowCoordinates(trip)}}{This method uses \code{trip} as an argument to return a qplot of the X-Y coordinates taken by the driver of the trip.}
 \item{\code{ShowVelocity(trip)}}{This method uses \code{trip} as an argument to return a qplot of the velocity (Moving average over 10 second interval) for the trip.}
 \item{\code{ShowAcceleration(trip)}}{This method uses \code{trip} as an argument to return a qplot of the acceleration over 10 seconds of the driver for that trip.}
 \item{\code{ShowBreaks(trip)}}{This method uses \code{trip} as an argument to return a qplot of the breaks over 4 seconds of the driver for that trip.}
 \item{\code{ShowSpeedBreaks(trip)}}{This method uses \code{trip} as an argument to return a qplot of the Speed Limit Breakages on Freeway for Driver}
 \item{\code{ShowQuantiles(trip)}}{This method uses \code{trip} as an argument to return TRUE if everything was OK.}
 \item{\code{CreateDriverSignature(trainClassMultipler = 5, trees = 100, shrinkage = .01)}}{This method returns a model for creating a driver unique signature.}
}
}
\keyword{data}

