% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{GridPlot}
\alias{GridPlot}
\title{A function to create grid plots}
\usage{
GridPlot(..., plotlist = NULL, cols)
}
\arguments{
\item{...}{Any arguments}

\item{plotlist}{List of plots}

\item{cols}{columns}
}
\value{
It will return a function of grid plot
}
\description{
A function to create grid plots
}

