% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{MovingAverage}
\alias{MovingAverage}
\title{A function to calculate the breaks using velocities obtained based on interval of 4 seconds.}
\usage{
MovingAverage(xv, n = 1, centered = FALSE)
}
\arguments{
\item{xv}{xv}

\item{n}{value=1}

\item{centered}{True or False}
}
\value{
It will return the breaks of the trip based on 4 second intervals.
}
\description{
Based on a certain trip specified, this function will be able to calculate the breaks using
the velocities of the trip that is based on interval of 4 seconds.
}
\examples{
# Breaks of the trip (based on velocity4)
 breaks(trip=sampleTrip)
}

