% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotggBreaks}
\alias{PlotggBreaks}
\title{A function to plots the breaks (Moving average over 4 second interval) for the trip.}
\usage{
PlotggBreaks(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
It will return a qplot of the brakes over 4 seconds of the driver for that trip.
}
\description{
Based on a certain trip specified, this function will be able to plot
the breaks  (Moving average over 4 second interval) occurred during the trip. The different colors show the different levels of breaks.
Note that breaks is the same as acceleration
}
\examples{
# Breaks plot of a trip taken by a driver
 PlotggBreaks(trip=sampleTrip)
}

