% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotggCoordinates}
\alias{PlotggCoordinates}
\title{A function to plot the X-Y coordinates taken by the driver of the trip.}
\usage{
PlotggCoordinates(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
It will return a qplot of the X-Y coordinates taken by the driver of the trip.
}
\description{
Based on a certain trip specified, this function will be able to plot
the XY Coordinates of a trip using ggplot scatter plot
}
\examples{
# X-Y coordinates plot of a trip taken by a driver
 PlotggCoordinates(trip=sampleTrip)
}

