% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotggSpeedLimitBreakage}
\alias{PlotggSpeedLimitBreakage}
\title{A function that shows the points of time the speed limit of 75KMPH was broken by the driver}
\usage{
PlotggSpeedLimitBreakage(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of
the trip}
}
\value{
It will return a qplot .
}
\description{
A function that shows the points of time the speed limit of 75KMPH was broken by the driver
}
\examples{
# Speed Limit Breaks plot of a trip taken by a driver
 PlotggSpeedLimitBreakage(trip=sampleTrip)
}

