% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotggTripNow}
\alias{PlotggTripNow}
\title{A function to plot a summary trip graph in a grid}
\usage{
PlotggTripNow(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
It will return a multi plot of the 4 features (X-Y coordinates, velocity, acceleration, breaks)
}
\description{
Based on a certain trip specified, this function will be able to plot
the 4 different features.
a) X-Y Coordinates
b) Velocity of the trip
c) Acceleration of the trip
d) Breaks of the trip
}
\examples{
# Plot of the 4 features (X-Y coordinates, velocity, acceleration, breaks)
 PlotggTripNow(trip=sampleTrip)
}

