% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotggVelocity}
\alias{PlotggVelocity}
\title{A function to plot the Velocity (Moving average over 10 second interval) for the trip}
\usage{
PlotggVelocity(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
It will return a qplot of the velocity (Moving average over 10 second interval) for the trip.
}
\description{
Based on a certain trip specified, this function will be able to plot
the velocity of the trip. The different colors show the status of the car,
to show whether the car is driving on the freeway or whether it is moving very slowly.
}
\examples{
# Velocity plot of a trip taken by a driver
 PlotggVelocity(trip=sampleTrip)
}

