% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PlotQuantiles}
\alias{PlotQuantiles}
\title{A function to plot the Quantiles for Velocity, Acceleration and Breaks}
\usage{
PlotQuantiles(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
The function will return TRUE if everything was OK
}
\description{
A function to plot the Quantiles for Velocity, Acceleration and Breaks
}
\examples{
# Velocity plot of a trip taken by a driver
 PlotQuantiles(trip=sampleTrip)
}

