% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{PrintAndStop}
\alias{PrintAndStop}
\title{A function to print a message and stop execution}
\usage{
PrintAndStop(msg)
}
\arguments{
\item{msg}{a string message}
}
\value{
It will stop execution
}
\description{
A function to print a message and stop execution
}

