% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sampleTrip.R
\docType{data}
\name{sampleTrip}
\alias{sampleTrip}
\title{Sample dataset used for examples}
\format{\preformatted{'data.frame':	863 obs. of  2 variables:
 $ x: num  0 18.6 36.1 53.7 70.1 ...
 $ y: num  0 -11.1 -21.9 -32.6 -42.8 -52.6 -62.3 -71.6 -80.4 -88.7 ...
}}
\usage{
sampleTrip
}
\description{
This data set lists the X-Y coordinates of a certain trip of a driver.
}
\details{
Sample dataset for Trip

This is a sample dataset for trip that is used in the examples.
}
\keyword{datasets}

