% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Telematics.R
\docType{class}
\name{Telematics}
\alias{Telematics}
\title{R6 class for Telematics}
\format{An \code{\link{R6Class}} generator object}
\source{
telematicsUtil.R
}
\usage{
Telematics
}
\description{
This is an R6 class for Telematics which contains methods to create a master summary file.
The mster summary file is a summarization of all trips by all drivers.
Each trip has about 66 features that are created by this class
}
\details{
This class needs a location for masterSummary.csv as part of constructors.
You can call load method on the class to load the summary file OR
You can call createMasterSummary to create a new one.
}
\section{Methods}{

\describe{
 \item{\code{load(trip)}}{This method uses \code{verbose = TRUE} as an argument to load the masterfile into memory.}
 \item{\code{summary()}}{This method prints a summary of the master file. }
 \item{\code{drivers()}}{This method returns the unique drivers loaded into memory.}
 \item{\code{CreateMasterSummary()}}{Creates the master summary csv file that has a summary of the various features of all the trips. \cr
             Method options and Defaults : \cr
              file.list ="fileList.csv" : The CSV file that has listing of all the trips\cr
              root.folder = ". : The root.folder where the directory structure is placed\cr
              subset.driver.count = 0 : the number of drivers you want in your master. IF 0 then all drivers considered. Note this can take a few hours as we have about half million trip. \cr
              cores = detectCores() : number of cores to run on\cr
             }
 \item{\code{SegmentDrivers()}}{This method runs a K-means clustering algorith to segment drivers in the clusters.
                               The user has to supply the numebr of segments. The return type if a k-means model}
 \item{\code{VisualizeSegments()}}{This method shows a clusterPlot of a previously create k-means model}

}
}
\keyword{data}

