% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{TravelDisList}
\alias{TravelDisList}
\title{A function to create a list of the travel Distance for the trip from start to end}
\usage{
TravelDisList(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
It will return a list of the travel distance in meters.
}
\description{
Based on a certain trip specified, this function will be able to create
a list of the travel distance for the trip from start to end
}
\examples{
# List of distance of a certain trip.
 TravelDisList(trip=sampleTrip)
}

