% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{TripDetails}
\alias{TripDetails}
\title{Main function to summarize a trip to a set of features represented by a row}
\usage{
TripDetails(trip)
}
\arguments{
\item{trip}{a dataframe with x and y columns representing the locations of the driver at each second of the trip}
}
\value{
a vector of 63 columns representing a trip
}
\description{
A Trip can be of varying timeframes. i.e a trip of 200 seconds will have 200 rows and trip of 1000 seconds #' will have 1000 rows.
This function normalize's all trips to 1 row of 63 features.
The features include
triptime, tripdistance(travel), tripStarttoend distance, 20 features for velocity quantiles,
20 features for acceleration quantiles, 20 features for breaks quantiles
}
\examples{
# Trip details of a certain trip
 TripDetails(trip=sampleTrip)
}

