% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/telematicsUtil.R
\name{velocity10}
\alias{velocity10}
\title{A function to calculate the velocities of the trip based on interval of 10 seconds.}
\usage{
velocity10(trip)
}
\arguments{
\item{trip}{The trip of the driver}
}
\value{
It will return the velocities of the trip in km/h based on 10 second intervals.
}
\description{
Based on a certain trip specified, this function will be able to calculate
the velocities of the trip based on interval of 10 seconds.
}
\examples{
# Velocities of the trip.
 velocity10(trip=sampleTrip)
}

